+++ app/Tabletop/Cartography/Exceptions/CartographyDenied.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace GreatMarketrealmTabletop\Tabletop\Cartography\Exceptions;
+
+use RuntimeException;
+
+defined('ABSPATH') || exit;
+
+final class CartographyDenied extends RuntimeException
+{
+}